LayoutTests/imported/w3c/web-platform-tests/payment-request/allowpaymentrequest/allowpaymentrequest-attribute-same-origin-bc-containers.https-expected.txt


FAIL iframe assert_equals: expected "Success" but got "Exception"
FAIL frame assert_equals: expected "Success" but got "Exception"
FAIL object assert_equals: expected "Success" but got "Exception"
FAIL embed assert_equals: expected "Success" but got "Exception"
